% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPL2008_2015}
\alias{EPL2008_2015}
\title{English Premier League match outcomes}
\format{A data frame with 3040 rows and 5 variables:
\describe{
\item{date}{date of the match}
\item{home_team}{name of the home team}
\item{away_team}{name of the away team}
\item{home_goals}{number of goals the home team scored}
\item{away_goals}{number of goals the away team scored}
...
}}
\source{
\url{https://www.kaggle.com/hugomathien/soccer}
}
\usage{
EPL2008_2015
}
\description{
All English Premier League match outcomes
during the seasons 2008/2009 to 2014/2015
}
\keyword{datasets}
